reorganize enemy script files
