+++ Assets/Characters/Enemies/Caster/Ghost/Prefabs/Ghost_Projectile.prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 8302880608021199780}
   - component: {fileID: -3873902634193119060}
   - component: {fileID: -6356404276773765587}
+  - component: {fileID: 6104979629877936911}
   m_Layer: 0
   m_Name: Ghost_Projectile
   m_TagString: Untagged
@@ -117,7 +118,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 8302880608021199803}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 62ed20598b21fec4cb6d5af20fd14a9c, type: 3}
+  m_Script: {fileID: 11500000, guid: cd3550219302deb4d81c3afe7bf71416, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   damage: 4
@@ -142,3 +143,17 @@ Rigidbody2D:
   m_SleepingMode: 1
   m_CollisionDetection: 0
   m_Constraints: 4
+--- !u!114 &6104979629877936911
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8302880608021199803}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 6ce4037087da99a4e81f1533e7f6d6d8, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  moveSpeed: 6
+  collisionOffset: 0.5
