+++ Assets/Characters/Enemies/WarCaster/Bringer Of Death/Prefabs/GraspingHand.prefab
@@ -117,7 +117,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 8405352401042620441}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 1426bacc885888e4c8cc7bcc1f60acb7, type: 3}
+  m_Script: {fileID: 11500000, guid: 2992508e18ce8834782cc8f57adb3252, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   damage: 4
