+++ Assets/Scripts/Characters/Enemies/Spells/Versions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 20fb2cc94dd38354a962ae8d03e94ab8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
