+++ Assets/Scripts/Characters/Enemies/Spells/Versions/ProjectileSpell.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 62ed20598b21fec4cb6d5af20fd14a9c
+guid: cd3550219302deb4d81c3afe7bf71416
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
