+++ Assets/Scripts/Characters/Enemies/Spells/Versions/TargetSpell.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1426bacc885888e4c8cc7bcc1f60acb7
+guid: 2992508e18ce8834782cc8f57adb3252
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
