+++ Assets/Scripts/Characters/Enemies/Versions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dd99b7026809c1240a75ab3d46ab7775
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
