+++ Assets/Scripts/Characters/Enemies/Versions/WarCasterEnemy.cs
@@ -1,9 +1,10 @@
 ﻿using Assets.Scripts.Helpers;
 using UnityEngine;
 
-namespace Assets.Scripts.Characters.Enemies
+namespace Assets.Scripts.Characters.Enemies.Versions
 {
-    public class MeleeEnemy : BaseEnemy
+    [RequireComponent(typeof(SpellBook))]
+    public class WarCasterEnemy : BaseEnemy
     {
         [SerializeField]
         public int Attack1Damage = 2;
@@ -14,11 +15,44 @@ namespace Assets.Scripts.Characters.Enemies
         [SerializeField]
         public float meleeAttackRange = 4f;
 
+        [SerializeField]
+        private float castingAttackRange = 8f;
+
+        private SpellBook spellBook;
+
+        private void Awake()
+        {
+            spellBook = GetComponent<SpellBook>();
+        }
+
         public override void TryAttack()
         {
-            if (AnimationIsPlaying(new string[] { "TakeHit", "Death", "Attack1", "Attack2" }))
+            if (AnimationIsPlaying(new string[] { "TakeHit", "Death", "Attack1", "Attack2", "Cast" }))
                 return;
 
+            TryCasting();
+            TryMeleeAttack();
+        }
+
+        public void TryCasting()
+        {
+            if (Vector3.Distance(playerController.transform.position, transform.position) > castingAttackRange)
+                return;
+
+            if (RandomHelper.GetRandom(98))
+                return;
+
+            animator.Play("Cast");
+
+            var randomSpell = RandomHelper.GetRandom(spellBook.GetSpells());
+
+            // Execute random spell
+            animator.Play("Cast");
+            randomSpell();
+        }
+
+        public void TryMeleeAttack()
+        {
             if (Vector3.Distance(playerController.transform.position, transform.position) > meleeAttackRange)
                 return;
